requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.23.0

# NLP and ML
transformers>=4.25.0
torch>=2.0.0
scikit-learn>=1.2.0
spacy>=3.5.0

# Vector Search
faiss-cpu>=1.7.0

# Graph Database
neo4j>=5.5.0

# Big Data
pyspark>=3.3.0

# Web Framework
fastapi>=0.95.0
uvicorn>=0.21.0
streamlit>=1.20.0

# PDF Processing
pdfplumber>=0.8.0
pytika>=1.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
python-json-logger>=2.0.0

# Utilities
requests>=2.28.0
tqdm>=4.65.0
python-dateutil>=2.8.0

# Development
pytest>=7.2.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Jupyter
jupyter>=1.0.0
ipython>=8.10.0
